% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TLkmean.R
\name{constract_k_mean}
\alias{constract_k_mean}
\title{There are cluster size limits and kmean that maintains the distribution of the original data}
\usage{
constract_k_mean(
  dataf,
  K,
  iter,
  min_num,
  max_num,
  b = 2,
  c = 3,
  norm = TRUE,
  d = 0.5
)
}
\arguments{
\item{dataf}{input data,a dataframe}

\item{K}{the number of the clusters}

\item{iter}{the number of iter}

\item{min_num}{Each cluster contains the fewest number of points}

\item{max_num}{The maximum number of points each cluster contains}

\item{b}{A multiple of the mean number of candidate points for each cluster center, which must be greater than 1, with a default value of 3}

\item{c}{Ensure that each point is assigned to at least c center points}

\item{norm}{Whether the metadata is positively distributed   bool}

\item{d}{The composition ratio of the loss function, the Euclidean distance of 1-d, and the probability error of d}
}
\value{
center and mix_matrix
}
\description{
There are cluster size limits and kmean that maintains the distribution of the original data
}
\examples{
load("TLkmean/data/data.RData")
out=constract_k_mean(dataf=dataf,K=500,iter=5,min_num=8,max_num=15,b=2,c=3,TRUE,0.6)

}
